mod for collisions

+++ kstring.h
@@ -37,7 +37,7 @@
 #ifndef KSTRING_T
 #define KSTRING_T kstring_t
 typedef struct __kstring_t {
-	size_t l, m;
+	uint32_t l, m;
 	char *s;
 } kstring_t;
 #endif
